Split hook utils into separate modules
