+++ PyInstaller/utils/hooks/__init__.py
@@ -8,13 +8,6 @@
 #-----------------------------------------------------------------------------
 
 
-# TODO: This module is getting a bit long in the tooth. Let's consider
-# splitting package-specific utility functions into new utility modules: e.g.,
-#
-# * "PyInstaller.util.hooks.django", containing all Django helpers.
-# * "PyInstaller.util.hooks.gi", containing all GObject-Introspection helpers.
-# * "PyInstaller.util.hooks.qt", containing all Qt helpers.
-
 import copy
 import glob
 import os
@@ -173,115 +166,8 @@ print(list(diff))
     return module_imports
 
 
-def qt_plugins_dir(namespace):
     """
-    Return list of paths searched for plugins
-
-    :param namespace: Import namespace, i.e., PyQt4, PyQt5, or PySide
-
-    :return: Plugin directory paths
     """
-    if namespace not in ['PyQt4', 'PyQt5', 'PySide']:
-        raise Exception('Invalid namespace: {0}'.format(namespace))
-    paths = eval_statement("""
-        from {0}.QtCore import QCoreApplication;
-        app = QCoreApplication([]);
-        # For Python 2 print would give <PyQt4.QtCore.QStringList
-        # object at 0x....>", so we need to convert each element separately
-        str = getattr(__builtins__, 'unicode', str);  # for Python 2
-        print([str(p) for p in app.libraryPaths()])
-        """.format(namespace))
-    if not paths:
-        raise Exception('Cannot find {0} plugin directories'.format(namespace))
-    else:
-        valid_paths = []
-        for path in paths:
-            if os.path.isdir(path):
-                valid_paths.append(str(path))  # must be 8-bit chars for one-file builds
-        qt_plugin_paths = valid_paths
-    if not qt_plugin_paths:
-        raise Exception("""
-            Cannot find existing {0} plugin directories
-            Paths checked: {1}
-            """.format(namespace, paths))
-    return qt_plugin_paths
-
-
-def qt_plugins_binaries(plugin_type, namespace):
-    """
-    Return list of dynamic libraries formatted for mod.binaries.
-
-    :param plugin_type: Plugin to look for
-    :param namespace: Import namespace, i.e., PyQt4, PyQt5 or PySide
-
-    :return: Plugin directory path corresponding to the given plugin_type
-    """
-    if namespace not in ['PyQt4', 'PyQt5', 'PySide']:
-        raise Exception('Invalid namespace: {0}'.format(namespace))
-    pdir = qt_plugins_dir(namespace=namespace)
-    files = []
-    for path in pdir:
-        files.extend(misc.dlls_in_dir(os.path.join(path, plugin_type)))
-
-    # Windows:
-    #
-    # dlls_in_dir() grabs all files ending with *.dll, *.so and *.dylib in a certain
-    # directory. On Windows this would grab debug copies of Qt plugins, which then
-    # causes PyInstaller to add a dependency on the Debug CRT __in addition__ to the
-    # release CRT.
-    #
-    # Since on Windows debug copies of Qt4 plugins end with "d4.dll" and Qt 5 plugins
-    # end with "d.dll" we filter them out of the list.
-    #
-    if is_win and (namespace in ['PyQt4', 'PySide']):
-        files = [f for f in files if not f.endswith("d4.dll")]
-    elif is_win and namespace is 'PyQt5':
-        files = [f for f in files if not f.endswith("d.dll")]
-
-    logger.debug('Found plugin files {0} for plugin \'{1}\''.format(files, plugin_type))
-    if namespace in ['PyQt4', 'PySide']:
-        plugin_dir = 'qt4_plugins'
-    else:
-        plugin_dir = 'qt5_plugins'
-    dest_dir = os.path.join(plugin_dir, plugin_type)
-    binaries = [
-        (f, dest_dir)
-        for f in files]
-    return binaries
-
-
-def qt_menu_nib_dir(namespace):
-    """
-    Return path to Qt resource dir qt_menu.nib on OSX only.
-
-    :param namespace: Import namespace, i.e., PyQt4, PyQt5, or PySide
-
-    :return: Directory containing qt_menu.nib for specified namespace
-    """
-    if namespace not in ['PyQt4', 'PyQt5', 'PySide']:
-        raise Exception('Invalid namespace: {0}'.format(namespace))
-    menu_dir = None
-
-    path = exec_statement("""
-    from {0}.QtCore import QLibraryInfo
-    path = QLibraryInfo.location(QLibraryInfo.LibrariesPath)
-    str = getattr(__builtins__, 'unicode', str)  # for Python 2
-    print(str(path))
-    """.format(namespace))
-    for location in [os.path.join(path, 'Resources'), os.path.join(path, 'QtGui.framework', 'Resources')]:
-        # Check directory existence
-        path = os.path.join(location, 'qt_menu.nib')
-        if os.path.exists(path):
-            menu_dir = path
-            logger.debug('Found qt_menu.nib for {0} at {1}'.format(namespace, path))
-            break
-    if not menu_dir:
-        raise Exception("""
-            Cannot find qt_menu.nib for {0}
-            Path checked: {1}
-            """.format(namespace, path))
-    return menu_dir
-
 def get_homebrew_path(formula = ''):
     '''Return the homebrew path to the requested formula, or the global prefix when
        called with no argument.  Returns the path as a string or None if not found.'''
@@ -306,143 +192,6 @@ def get_homebrew_path(formula = ''):
         return path
     else:
         return None
-
-def get_qmake_path(version = ''):
-    '''
-    Try to find the path to qmake with version given by the argument
-    as a string.
-    '''
-    import subprocess
-
-    # Use QT[45]DIR if specified in the environment
-    if 'QT5DIR' in os.environ and version[0] == '5':
-        logger.debug('Using $QT5DIR/bin as qmake path')
-        return os.path.join(os.environ['QT5DIR'],'bin','qmake')
-    if 'QT4DIR' in os.environ and version[0] == '4':
-        logger.debug('Using $QT4DIR/bin as qmake path')
-        return os.path.join(os.environ['QT4DIR'],'bin','qmake')
-
-    # try the default $PATH
-    dirs = ['']
-
-    # try homebrew paths
-    for formula in ('qt','qt5'):
-        homebrewqtpath = get_homebrew_path(formula)
-        if homebrewqtpath:
-            dirs.append(homebrewqtpath)
-
-    for dir in dirs:
-        try:
-            qmake = os.path.join(dir, 'qmake')
-            versionstring = subprocess.check_output([qmake, '-query',
-                                                     'QT_VERSION']).strip()
-            if is_py3:
-                # version string is probably just ASCII
-                versionstring = versionstring.decode('utf8')
-            if versionstring.find(version) == 0:
-                logger.debug('Found qmake version "%s" at "%s".'
-                             % (versionstring, qmake))
-                return qmake
-        except (OSError, subprocess.CalledProcessError):
-            pass
-    logger.debug('Could not find qmake matching version "%s".' % version)
-    return None
-
-
-def qt5_qml_dir():
-    qmake = get_qmake_path('5')
-    if qmake is None:
-        qmldir = ''
-        logger.error('Could not find qmake version 5.x, make sure PATH is '
-                     'set correctly or try setting QT5DIR.')
-    else:
-        qmldir = compat.exec_command(qmake, "-query", "QT_INSTALL_QML").strip()
-    if len(qmldir) == 0:
-        logger.error('Cannot find QT_INSTALL_QML directory, "qmake -query '
-                        + 'QT_INSTALL_QML" returned nothing')
-    elif not os.path.exists(qmldir):
-        logger.error("Directory QT_INSTALL_QML: %s doesn't exist" % qmldir)
-
-    # 'qmake -query' uses / as the path separator, even on Windows
-    qmldir = os.path.normpath(qmldir)
-    return qmldir
-
-def qt5_qml_data(dir):
-    """Return Qml library dir formatted for data"""
-    qmldir = qt5_qml_dir()
-    return (os.path.join(qmldir, dir), 'qml')
-
-def qt5_qml_plugins_binaries(dir):
-    """Return list of dynamic libraries formatted for mod.binaries."""
-    binaries = []
-    qmldir = qt5_qml_dir()
-    files = misc.dlls_in_subdirs(os.path.join(qmldir, dir))
-    for f in files:
-        relpath = os.path.relpath(f, qmldir)
-        instdir, file = os.path.split(relpath)
-        instdir = os.path.join("qml", instdir)
-        logger.debug("qt5_qml_plugins_binaries installing %s in %s"
-                     % (f, instdir) )
-        binaries.append((f, instdir))
-    return binaries
-
-
-def django_dottedstring_imports(django_root_dir):
-    """
-    Get all the necessary Django modules specified in settings.py.
-
-    In the settings.py the modules are specified in several variables
-    as strings.
-    """
-    pths = []
-    # Extend PYTHONPATH with parent dir of django_root_dir.
-    pths.append(misc.get_path_to_toplevel_modules(django_root_dir))
-    # Extend PYTHONPATH with django_root_dir.
-    # Many times Django users do not specify absolute imports in the settings module.
-    pths.append(django_root_dir)
-
-    package_name = os.path.basename(django_root_dir) + '.settings'
-    env = {'DJANGO_SETTINGS_MODULE': package_name,
-           'PYTHONPATH': os.pathsep.join(pths)}
-    ret = eval_script('django_import_finder.py', env=env)
-
-    return ret
-
-
-def django_find_root_dir():
-    """
-    Return path to directory (top-level Python package) that contains main django
-    files. Return None if no directory was detected.
-
-    Main Django project directory contain files like '__init__.py', 'settings.py'
-    and 'url.py'.
-
-    In Django 1.4+ the script 'manage.py' is not in the directory with 'settings.py'
-    but usually one level up. We need to detect this special case too.
-    """
-    # 'PyInstaller.config' cannot be imported as other top-level modules.
-    from ...config import CONF
-    # Get the directory with manage.py. Manage.py is supplied to PyInstaller as the
-    # first main executable script.
-    manage_py = CONF['main_script']
-    manage_dir = os.path.dirname(os.path.abspath(manage_py))
-
-    # Get the Django root directory. The directory that contains settings.py and url.py.
-    # It could be the directory containig manage.py or any of its subdirectories.
-    settings_dir = None
-    files = set(os.listdir(manage_dir))
-    if 'settings.py' in files and 'urls.py' in files:
-        settings_dir = manage_dir
-    else:
-        for f in files:
-            if os.path.isdir(os.path.join(manage_dir, f)):
-                subfiles = os.listdir(os.path.join(manage_dir, f))
-                # Subdirectory contains critical files.
-                if 'settings.py' in subfiles and 'urls.py' in subfiles:
-                    settings_dir = os.path.join(manage_dir, f)
-                    break  # Find the first directory.
-
-    return settings_dir
 
 
 # TODO Move to "hooks/hook-OpenGL.py", the only place where this is called.
@@ -586,50 +335,6 @@ print(p.__file__)
         if not attr.strip():
             raise ImportError
     return attr
-
-
-# TODO: Rename to get_pywin32_module_dll_path() and move to a new
-# "PyInstaller.utils.hooks.win32" module.
-# NOTE: This function requires PyInstaller to be on the default "sys.path" for
-# the called Python process. Running py.test changes the working dir to a temp
-# dir, so PyInstaller should be installed via either "setup.py install" or
-# "setup.py develop" before running py.test.
-def get_pywin32_module_file_attribute(module_name):
-    """
-    Get the absolute path of the PyWin32 DLL specific to the PyWin32 module
-    with the passed name.
-
-    On import, each PyWin32 module:
-
-    * Imports a DLL specific to that module.
-    * Overwrites the values of all module attributes with values specific to
-      that DLL. This includes that module's `__file__` attribute, which then
-      provides the absolute path of that DLL.
-
-    This function safely imports that module in a PyWin32-aware subprocess and
-    returns the value of that module's `__file__` attribute.
-
-    Parameters
-    ----------
-    module_name : str
-        Fully-qualified name of that module.
-
-    Returns
-    ----------
-    str
-        Absolute path of that DLL.
-
-    See Also
-    ----------
-    `PyInstaller.utils.win32.winutils.import_pywin32_module()`
-        For further details.
-    """
-    statement = """
-from PyInstaller.utils.win32 import winutils
-module = winutils.import_pywin32_module('%s')
-print(module.__file__)
-"""
-    return exec_statement(statement % module_name)
 
 
 def is_module_satisfies(
@@ -1056,227 +761,6 @@ def relpath_to_config_or_make(filename):
     return os.path.relpath(os.path.dirname(filename), prefix)
 
 
-def get_typelibs(module, version):
-    '''deprecated; only here for backwards compat'''
-    logger.warn("get_typelibs is deprecated, use get_gi_typelibs instead")
-    return get_gi_typelibs(module, version)[1]
-
-def get_gi_libdir(module, version):
-    statement = """
-import gi
-gi.require_version("GIRepository", "2.0")
-from gi.repository import GIRepository
-repo = GIRepository.Repository.get_default()
-module, version = (%r, %r)
-repo.require(module, version,
-             GIRepository.RepositoryLoadFlags.IREPOSITORY_LOAD_FLAG_LAZY)
-print(repo.get_shared_library(module))
-"""
-    statement %= (module, version)
-    libs = exec_statement(statement).split(',')
-    for lib in libs:
-        path = findSystemLibrary(lib.strip())
-        return os.path.normpath(os.path.dirname(path))
-
-    raise ValueError("Could not find libdir for %s-%s" % (module, version))
-
-def get_gi_typelibs(module, version):
-    """
-    Returns a tuple of (binaries, datas, hiddenimports) to be used by PyGObject
-    related hooks. Searches for and adds dependencies recursively.
-
-    :param module: GI module name, as passed to 'gi.require_version()'
-    :param version: GI module version, as passed to 'gi.require_version()'
-    """
-    datas = []
-    binaries = []
-    hiddenimports = []
-
-    statement = """
-import gi
-gi.require_version("GIRepository", "2.0")
-from gi.repository import GIRepository
-repo = GIRepository.Repository.get_default()
-module, version = (%r, %r)
-repo.require(module, version,
-             GIRepository.RepositoryLoadFlags.IREPOSITORY_LOAD_FLAG_LAZY)
-get_deps = getattr(repo, 'get_immediate_dependencies', None)
-if not get_deps:
-    get_deps = repo.get_dependencies
-print({'sharedlib': repo.get_shared_library(module),
-       'typelib': repo.get_typelib_path(module),
-       'deps': get_deps(module) or []})
-"""
-    statement %= (module, version)
-    typelibs_data = eval_statement(statement)
-    if not typelibs_data:
-        logger.error("gi repository 'GIRepository 2.0' not found. "
-                     "Please make sure libgirepository-gir2.0 resp. "
-                     "lib64girepository-gir2.0 is installed.")
-        # :todo: should we raise a SystemError here?
-    else:
-        logger.debug("Adding files for %s %s", module, version)
-
-        if typelibs_data['sharedlib']:
-            for lib in typelibs_data['sharedlib'].split(','):
-                path = findSystemLibrary(lib.strip())
-                if path:
-                    logger.debug('Found shared library %s at %s', lib, path)
-                    binaries.append((path, ''))
-
-        d = gir_library_path_fix(typelibs_data['typelib'])
-        if d:
-            logger.debug('Found gir typelib at %s', d)
-            datas.append(d)
-
-        hiddenimports += collect_submodules('gi.overrides',
-                                            pattern='.' + module, endswith=True)
-
-        ## Load dependencies recursively
-        for dep in typelibs_data['deps']:
-            m, _ = dep.rsplit('-', 1)
-            hiddenimports += ['gi.repository.%s' % m]
-
-    return binaries, datas, hiddenimports
-
-
-def gir_library_path_fix(path):
-    import subprocess
-    # 'PyInstaller.config' cannot be imported as other top-level modules.
-    from ...config import CONF
-
-    path = os.path.abspath(path)
-    common_path = os.path.commonprefix([compat.base_prefix, path])
-    gir_path = os.path.join(common_path, 'share', 'gir-1.0')
-
-    typelib_name = os.path.basename(path)
-    gir_name = os.path.splitext(typelib_name)[0] + '.gir'
-
-    gir_file = os.path.join(gir_path, gir_name)
-
-    if is_darwin:
-        if not os.path.exists(gir_path):
-            logger.error('Unable to find gir directory: %s.\n'
-                         'Try installing your platforms gobject-introspection '
-                         'package.', gir_path)
-            return None
-        if not os.path.exists(gir_file):
-            logger.error('Unable to find gir file: %s.\n'
-                         'Try installing your platforms gobject-introspection '
-                         'package.', gir_file)
-            return None
-
-        with open(gir_file, 'r') as f:
-            lines = f.readlines()
-        with open(os.path.join(CONF['workpath'], gir_name), 'w') as f:
-            for line in lines:
-                if 'shared-library' in line:
-                    split = re.split('(=)', line)
-                    files = re.split('(["|,])', split[2])
-                    for count, item in enumerate(files):
-                        if 'lib' in item:
-                            files[count] = '@loader_path/' + os.path.basename(item)
-                    line = ''.join(split[0:2]) + ''.join(files)
-                f.write(line)
-
-        # g-ir-compiler expects a file so we cannot just pipe the fixed file to it.
-        command = subprocess.Popen(('g-ir-compiler', os.path.join(CONF['workpath'], gir_name),
-                                    '-o', os.path.join(CONF['workpath'], typelib_name)))
-        command.wait()
-
-        return os.path.join(CONF['workpath'], typelib_name), 'gi_typelibs'
-    else:
-        return (path, 'gi_typelibs')
-
-def get_glib_system_data_dirs():
-    statement = """
-import gi
-gi.require_version('GLib', '2.0')
-from gi.repository import GLib
-print(GLib.get_system_data_dirs())
-"""
-    data_dirs = eval_statement(statement)
-    if not data_dirs:
-        logger.error("gi repository 'GIRepository 2.0' not found. "
-                     "Please make sure libgirepository-gir2.0 resp. "
-                     "lib64girepository-gir2.0 is installed.")
-        # :todo: should we raise a SystemError here?
-    return data_dirs
-
-def get_glib_sysconf_dirs():
-    '''Tries to return the sysconf directories, eg /etc'''
-
-    if is_win:
-        # On windows, if you look at gtkwin32.c, sysconfdir is actually
-        # relative to the location of the GTK DLL. Since that's what
-        # we're actually interested in (not the user path), we have to
-        # do that the hard way'''
-        return [os.path.join(get_gi_libdir('GLib', '2.0'), 'etc')]
-
-    statement = """
-import gi
-gi.require_version('GLib', '2.0')
-from gi.repository import GLib
-print(GLib.get_system_config_dirs())
-"""
-    data_dirs = eval_statement(statement)
-    if not data_dirs:
-        logger.error("gi repository 'GIRepository 2.0' not found. "
-                     "Please make sure libgirepository-gir2.0 resp. "
-                     "lib64girepository-gir2.0 is installed.")
-        # :todo: should we raise a SystemError here?
-    return data_dirs
-
-def collect_glib_share_files(*path):
-    '''path is relative to the system data directory (eg, /usr/share)'''
-    glib_data_dirs = get_glib_system_data_dirs()
-    if glib_data_dirs == None:
-        return []
-
-    destdir = os.path.join('share', *path[:-1])
-
-    # TODO: will this return too much?
-    collected = []
-    for data_dir in glib_data_dirs:
-        p = os.path.join(data_dir, *path)
-        collected += collect_system_data_files(p, destdir=destdir, include_py_files=False)
-
-    return collected
-
-def collect_glib_etc_files(*path):
-    '''path is relative to the system config directory (eg, /etc)'''
-    glib_config_dirs = get_glib_sysconf_dirs()
-    if glib_config_dirs == None:
-        return []
-
-    destdir = os.path.join('etc', *path[:-1])
-
-    # TODO: will this return too much?
-    collected = []
-    for config_dir in glib_config_dirs:
-        p = os.path.join(config_dir, *path)
-        collected += collect_system_data_files(p, destdir=destdir, include_py_files=False)
-
-    return collected
-
-_glib_translations = None
-
-def collect_glib_translations(prog):
-    """
-    Returns a list of translations in the system locale directory whose
-    names equal prog.mo
-    """
-
-    global _glib_translations
-    if _glib_translations is None:
-        _glib_translations = collect_glib_share_files('locale')
-
-    names = [os.sep + prog + '.mo',
-             os.sep + prog + '.po']
-    namelen = len(names[0])
-
-    return [(src, dst) for src, dst in _glib_translations if src[-namelen:] in names]
-
 def copy_metadata(package_name):
     """
     This function returns a list to be assigned to the ``datas`` global
@@ -1343,3 +827,9 @@ def copy_metadata(package_name):
       package_name, metadata_dir, dest_dir))
 
     return [ (metadata_dir, dest_dir) ]
+
+# These imports need to be here due to these modules recursively importing this module.
+from .django import *
+from .gi import *
+from .qt import *
+from .win32 import *
